support p2p between reth nodes

+++ example/test-local-chain/from_docker/stop_and_clean.sh
@@ -5,4 +5,11 @@ docker compose -f compose-mala.yaml down
 docker compose down
 
 rm -fr ./nodes
-rm -fr ./rethdata
+# Preserving reth0 discovery-secret files and cleaning rethdata...
+# reth0 is bootnode
+if [ -d "./rethdata/0" ]; then
+    mv ./rethdata/0/discovery-secret ./discovery-secret-0
+    rm -fr ./rethdata
+    mkdir -p ./rethdata/0
+    mv ./discovery-secret-0 ./rethdata/0/discovery-secret
+fi
